refactor: qrcode_screen is now broken up into a proper folder structure
