+++ frontend/lib/screens/qr_code/qr_code_constants.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+abstract final class QRCodeConstants {
+  // ─── BRANDING COLORS ──────────────────────────────────────────────────────
+
+  /// Primary maroon used for borders, icons, and accent elements.
+  static const Color maroon = Color(0xFF7A1D1D);
+
+  /// Darker maroon used for buttons, dialog headers, and filled surfaces.
+  static const Color maroonDark = Color(0xFF5D1414);
+
+  /// Light grey used as the AppBar background colour.
+  static const Color headerGrey = Color(0xFFF5F5F5);
+}
